Retain native SQLite DiskANN canonical sources (#171)
* Add guarded native canonical origin records

* Retain native DiskANN canonical tensors and mutation origins

+++ crates/uqa-storage-sqlite/src/mvcc/native/tests/graph_lookup/path_migration.rs
@@ -150,7 +150,7 @@ fn native_path_lookup_upgrade_preserves_sources_and_old_selectors_in_every_file_
             }
             assert_eq!(
                 dump(&connection, "SELECT format FROM _uqa_mvcc_native_format"),
-                vec![vec![rusqlite::types::Value::Integer(10)]]
+                vec![vec![rusqlite::types::Value::Integer(11)]]
             );
             assert_eq!(
                 SQLiteRecordStore::for_native(&connection, &control)
